fix: fold bot round-5 — oracle portability, BOM strip, skip placement
Three of five advisory findings verified real (two HIGHs refuted with
compile evidence — see PR thread):

- The coverage oracle used a machine-local temp dir (dir="/data/tmp");
  now the system default, so it runs anywhere including upstream CI.
- The .fs sniffer strips a leading UTF-8 BOM before line-anchoring —
  Windows-authored F# commonly leads with one, and a single-marker-line
  file would otherwise miss its only strong F# marker. Regression test.
- test_missing_grammar_is_reported_not_raised moved out of the
  importorskip'd module: it FAKES the ImportError, so it must run
  precisely on machines where the grammar is actually absent.
